+++ source_files/edge/vm_hud.cc
@@ -427,7 +427,7 @@ static void HD_draw_text(coal::vm_c *vm, int argc)
 
 	double *size = vm->AccessParam(3);
 
-	HUD_DrawText(x, y, str, size ? (int)*size : 0);
+	HUD_DrawText(x, y, str, size ? *size : 0);
 }
 
 
@@ -474,7 +474,7 @@ static void HD_draw_num2(coal::vm_c *vm, int argc)
 	}
 
 	HUD_SetAlignment(+1, -1);
-	HUD_DrawText(x, y, pos, size ? (int)*size : 0);
+	HUD_DrawText(x, y, pos, size ? *size : 0);
 	HUD_SetAlignment();
 }
 
@@ -524,12 +524,12 @@ static void HD_draw_number(coal::vm_c *vm, int argc)
 
 	if (align_right == 0)
 	{
-		HUD_DrawText(x, y, pos, size ? (int)*size : 0);
+		HUD_DrawText(x, y, pos, size ? *size : 0);
 	}
 	else
 	{
 		HUD_SetAlignment(+1, -1);
-		HUD_DrawText(x, y, pos, size ? (int)*size : 0);
+		HUD_DrawText(x, y, pos, size ? *size : 0);
 		HUD_SetAlignment();
 	}
 }
